Added experience 2 with dora world model.
In it, a class dependent and and a class independent probability distribution is learned on "known" density.
And a simple not-class-aware is used to model the unknown.

This is interesting as its makes it possible to model using a graphical models.

TODO: distributions are very strickyly calculated atm and they return 0 for values never seen.
That should be smoothed out.